Added global rgb and pixel buffer object to save memory
For functions that need to store to a rgb or pixel buffer object, creating a
new static rgb, and especially pixel buffer object is a wasteful use of memory.

Since at no time will two different routines have to access a rgb or pixelbuffer
object, a single global rgb and pixel buffer object will suffice.

The pixel buffer object MUST be tested for initialization, and if required, be
initialized,  by every function that wishes to make use of it. This is
to not waste memory for builds that do not use functions that make use of
the global pixel buffer.

Signed-off-by: Patrick Pedersen <ctx.xda@gmail.com>